# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Dash - Online Food Ordering System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateApplicationBodyRole } from "./createApplicationBodyRole";

export interface CreateApplicationBody {
  fullName: string;
  email: string;
  phone: string;
  role: CreateApplicationBodyRole;
  portfolioLink?: string;
  message?: string;
}