# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Dash - Online Food Ordering System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItem } from "./createOrderItem";

export interface CreateOrderBody {
  restaurantId: number;
  items: CreateOrderItem[];
  deliveryAddress: string;
  customerName: string;
  customerPhone: string;
}