# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Dash - Online Food Ordering System API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryOrdersByStatus } from "./dashboardSummaryOrdersByStatus";
import type { Order } from "./order";

export interface DashboardSummary {
  totalRestaurants: number;
  totalMenuItems: number;
  totalOrders: number;
  totalApplications: number;
  recentOrders: Order[];
  ordersByStatus: DashboardSummaryOrdersByStatus;
}