# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Dash - Online Food Ordering System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  restaurantId: number;
  restaurantName: string;
  items: OrderItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  status: OrderStatus;
  deliveryAddress: string;
  customerName: string;
  customerPhone: string;
  estimatedDelivery: string;
  createdAt: Date;
}