# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Dash - Online Food Ordering System API
 * OpenAPI spec version: 0.1.0
 */
import type { RestaurantCategory } from "./restaurantCategory";

export interface Restaurant {
  id: number;
  name: string;
  description: string;
  image: string;
  category: RestaurantCategory;
  rating: number;
  deliveryTimeMin: number;
  deliveryTimeMax: number;
  deliveryFee: number;
  tags: string[];
  isOpen: boolean;
  address: string;
  createdAt: Date;
}