# lib/api-zod/src/generated/types/updateRestaurantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Dash - Online Food Ordering System API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateRestaurantBodyCategory } from "./updateRestaurantBodyCategory";

export interface UpdateRestaurantBody {
  name?: string;
  description?: string;
  image?: string;
  category?: UpdateRestaurantBodyCategory;
  rating?: number;
  deliveryTimeMin?: number;
  deliveryTimeMax?: number;
  deliveryFee?: number;
  tags?: string[];
  isOpen?: boolean;
  address?: string;
}